+++ plone/app/upgrade/v51/final.py
@@ -1,6 +1,9 @@
 # -*- coding: utf-8 -*-
+from plone.app.upgrade.utils import cleanUpSkinsTool
+from plone.registry.interfaces import IRegistry
 from Products.CMFCore.utils import getToolByName
 from zExceptions import BadRequest
+from zope.component import getUtility
 
 import logging
 
@@ -38,3 +41,24 @@ def fix_i18n_domain(context):
                 'Action object/%s does not have an i18n_domain property',
                 action_id,
             )
+
+
+def remove_highlightsearchterms(context):
+    portal = getToolByName(context, 'portal_url').getPortalObject()
+    cleanUpSkinsTool(portal)
+
+    registry = getUtility(IRegistry)
+    record = 'plone.bundles/plone-legacy.resources'
+    resources = registry.records[record]
+    if u'jquery-highlightsearchterms' in resources.value:
+        resources.value.remove(u'jquery-highlightsearchterms')
+
+
+def remove_old_PAE_rescources(context):  # noqa
+    """FORCE remove old p.a.event resources"""
+    registry = getUtility(IRegistry)
+    resources = registry.records['plone.bundles/plone-legacy.resources']
+    if u'resource-plone-app-event-event-js' in resources.value:
+        resources.value.remove('resource-plone-app-event-event-js')
+    if u'resource-plone-app-event-event-css' in resources.value:
+        resources.value.remove('resource-plone-app-event-event-css')